+++ Source/WinCopies.Util.Desktop.Shared/Commands/ApplicationCommands.cs
@@ -33,51 +33,51 @@ namespace WinCopies.Commands
         /// <summary>
         /// Gets the <b>NewTab</b> command.
         /// </summary>
-        public static RoutedUICommand NewTab { get; } = new RoutedUICommand(WinCopies.
+        public static RoutedUICommand NewTab { get; } = new RoutedUICommand(
 #if !WinCopies3
-            Util.
+            WinCopies.Util.
 #endif
             Desktop.Resources.Commands.WPF.ApplicationCommands.NewTab, nameof(NewTab), typeof(ApplicationCommands), new InputGestureCollection() { new KeyGesture(Key.T, ModifierKeys.Control) });
 
         /// <summary>
         /// Gets the <b>NewWindow</b> command.
         /// </summary>
-        public static RoutedUICommand NewWindow { get; } = new RoutedUICommand(WinCopies.
+        public static RoutedUICommand NewWindow { get; } = new RoutedUICommand(
 #if !WinCopies3
-            Util.
+            WinCopies.Util.
 #endif
             Desktop.Resources.Commands.WPF.ApplicationCommands.NewWindow, nameof(NewWindow), typeof(ApplicationCommands), new InputGestureCollection() { new KeyGesture(Key.W, ModifierKeys.Alt) });
 
         /// <summary>
         /// Gets the <b>NewWindowInNewInstance</b> command.
         /// </summary>
-        public static RoutedUICommand NewWindowInNewInstance { get; } = new RoutedUICommand(WinCopies.
+        public static RoutedUICommand NewWindowInNewInstance { get; } = new RoutedUICommand(
 #if !WinCopies3
-            Util.
+            WinCopies.Util.
 #endif
             Desktop.Resources.Commands.WPF.ApplicationCommands.NewWindowInNewInstance, nameof(NewWindowInNewInstance), typeof(ApplicationCommands), new InputGestureCollection() { new KeyGesture(Key.W, ModifierKeys.Control | ModifierKeys.Shift) });
 
         /// <summary>
         /// Gets the <b>CloseTab</b> command.
         /// </summary>
-        public static RoutedUICommand CloseTab { get; } = new RoutedUICommand(WinCopies.
+        public static RoutedUICommand CloseTab { get; } = new RoutedUICommand(
 #if !WinCopies3
-            Util.
+            WinCopies.Util.
 #endif
             Desktop.Resources.Commands.WPF.ApplicationCommands.CloseTab, nameof(CloseTab), typeof(ApplicationCommands), new InputGestureCollection() { new KeyGesture(Key.W, ModifierKeys.Control) });
 
-        public static RoutedUICommand CloseOtherTabs { get; } = new RoutedUICommand(WinCopies.
+        public static RoutedUICommand CloseOtherTabs { get; } = new RoutedUICommand(
 #if !WinCopies3
-            Util.
+            WinCopies.Util.
 #endif
             Desktop.Resources.Commands.WPF.ApplicationCommands.CloseOtherTabs, nameof(CloseOtherTabs), typeof(ApplicationCommands), new InputGestureCollection() { new KeyGesture(Key.W, ModifierKeys.Control | ModifierKeys.Alt) });
 
         /// <summary>
         /// Gets the <b>CloseAllTabs</b> command.
         /// </summary>
-        public static RoutedUICommand CloseAllTabs { get; } = new RoutedUICommand(WinCopies.
+        public static RoutedUICommand CloseAllTabs { get; } = new RoutedUICommand(
 #if !WinCopies3
-            Util.
+            WinCopies.Util.
 #endif
             Desktop.Resources.Commands.WPF.ApplicationCommands.CloseAllTabs, nameof(CloseAllTabs), typeof(ApplicationCommands), new InputGestureCollection() { new KeyGesture(Key.W, ModifierKeys.Control | ModifierKeys.Shift) });
 
@@ -89,45 +89,45 @@ namespace WinCopies.Commands
         public static RoutedUICommand CloseWindow { get; } = new RoutedUICommand(Desktop.Resources.Commands.WPF.ApplicationCommands.CloseWindow, nameof(CloseWindow), typeof(ApplicationCommands), new InputGestureCollection() { new KeyGesture(Key.F4, ModifierKeys.Alt) });
 #endif
 
-        public static RoutedUICommand Quit { get; } = new RoutedUICommand(WinCopies.
+        public static RoutedUICommand Quit { get; } = new RoutedUICommand(
 #if !WinCopies3
-            Util.
+            WinCopies.Util.
 #endif
             Desktop.Resources.Commands.WPF.ApplicationCommands.Quit, nameof(Quit), typeof(ApplicationCommands), new InputGestureCollection() { new KeyGesture(Key.Q, ModifierKeys.Control) });
 
         /// <summary>
         /// Gets the <b>DeselectAll</b> command.
         /// </summary>
-        public static RoutedUICommand DeselectAll { get; } = new RoutedUICommand(WinCopies.
+        public static RoutedUICommand DeselectAll { get; } = new RoutedUICommand(
 #if !WinCopies3
-            Util.
+            WinCopies.Util.
 #endif
             Desktop.Resources.Commands.WPF.ApplicationCommands.DeselectAll, nameof(DeselectAll), typeof(ApplicationCommands));
 
         /// <summary>
         /// Gets the <b>ReverseSelection</b> command.
         /// </summary>
-        public static RoutedUICommand ReverseSelection { get; } = new RoutedUICommand(WinCopies.
+        public static RoutedUICommand ReverseSelection { get; } = new RoutedUICommand(
 #if !WinCopies3
-            Util.
+            WinCopies.Util.
 #endif
             Desktop.Resources.Commands.WPF.ApplicationCommands.ReverseSelection, nameof(ReverseSelection), typeof(ApplicationCommands));
 
-        public static RoutedUICommand Reset { get; } = new RoutedUICommand(WinCopies.
+        public static RoutedUICommand Reset { get; } = new RoutedUICommand(
 #if !WinCopies3
-            Util.
+            WinCopies.Util.
 #endif
             Desktop.Resources.Commands.WPF.ApplicationCommands.Reset, nameof(Reset), typeof(ApplicationCommands));
 
-        public static RoutedUICommand Empty { get; } = new RoutedUICommand(WinCopies.
+        public static RoutedUICommand Empty { get; } = new RoutedUICommand(
 #if !WinCopies3
-            Util.
+            WinCopies.Util.
 #endif
             Desktop.Resources.Commands.WPF.ApplicationCommands.Empty, nameof(Empty), typeof(ApplicationCommands));
 
-        public static RoutedUICommand DeletePermanently { get; } = new RoutedUICommand(WinCopies.
+        public static RoutedUICommand DeletePermanently { get; } = new RoutedUICommand(
 #if !WinCopies3
-            Util.
+            WinCopies.Util.
 #endif
             Desktop.Resources.Commands.WPF.ApplicationCommands.DeletePermanently, nameof(DeletePermanently), typeof(ApplicationCommands), new InputGestureCollection() { new System.Windows.Input.KeyGesture(Key.Delete, ModifierKeys.Alt) });
     }